# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Pemuda Surau Al-Abqori API
  version: 1.0.0
  description: |
    REST API for the Surau Al-Abqori youth programme: prayer attendance
    (face-verified + geofenced), Quran activity, merits, programs and the
    public dashboard.

    ## Authentication
    Most endpoints require a JWT obtained from `POST /api/auth/verify-otp`.
    Send it as `Authorization: Bearer <token>`.

    ## Authorisation
    Beyond "logged in", endpoints are gated by **capabilities** rather than raw
    roles (see `server/src/utils/roles.js`). Admins implicitly hold every
    capability. The `x-capability` extension on each operation records the
    requirement.

    ## Geofencing
    Check-in, activity recording and new registrations are only accepted within
    a radius of the surau. Coordinates are supplied by the client.
servers:
  - url: /api
    description: Same origin (proxied by Nginx / the Vite dev server)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Authentication required
        code:
          type: string
          description: Optional machine-readable code (e.g. REGISTRATION_TOO_FAR).
          example: REGISTRATION_TOO_FAR
        retryAfterMinutes:
          type: integer
          description: Present on 429 responses (e.g. the Quran log cooldown).
...
    Descriptor:
      type: array
      description: 128-float face descriptor produced by face-api.js in the browser.
      items: { type: number }
      minItems: 128
      maxItems: 128
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Authenticated but lacking the required capability.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /activity/merits:
    get:
      tags: [Activity]
      summary: All merits (staff)
      x-capability: manageMerits
      responses:
        '200': { description: OK }
    post:
      tags: [Activity]
      summary: Award merits
      description: |
        Requires `manageMerits`. Must be within the geofence, and the supplied
        face descriptor must match the member being awarded.
      x-capability: manageMerits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, points, descriptor, latitude, longitude]
              properties:
                userId: { type: string, format: uuid }
                points: { type: integer, minimum: 1, maximum: 100 }
                reason: { type: string, nullable: true }
                descriptor: { $ref: '#/components/schemas/Descriptor' }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '201': { description: Awarded }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /activity/merits/{id}:
    delete:
      tags: [Activity]
      summary: Revoke a merit
      x-capability: manageMerits
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Revoked }
...
  /activity/quran:
    get:
      tags: [Activity]
      summary: All Quran logs (staff)
      x-capability: manageQuran
      responses:
        '200': { description: OK }
    post:
      tags: [Activity]
      summary: Log Quran activity
      description: |
        Every log must be recorded at the surau (geofence). At most one log per
        member per cooldown window (default 60 minutes) — a 429 is returned with
        `retryAfterMinutes`. Teachers/admins may log for any member (a matching
        face descriptor is then required); members may log for themselves or a
        dependent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, latitude, longitude]
              properties:
                kind: { type: string, enum: [recitation, memorization] }
                surah: { type: string, nullable: true }
                juz: { type: integer, minimum: 1, maximum: 30, nullable: true }
                pages: { type: integer, minimum: 1, nullable: true }
                note: { type: string, nullable: true }
                forUserId: { type: string, format: uuid, nullable: true }
                descriptor: { $ref: '#/components/schemas/Descriptor' }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '201': { description: Logged }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429':
          description: 'Cooldown active (code: QURAN_COOLDOWN).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /activity/quran/{id}:
    delete:
      tags: [Activity]
      summary: Delete a Quran log
      description: Own, a dependent's, or any member for teachers/admins.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Deleted }

  # ------------------------------------------------------------- Dashboard
...